Rename compat to bridge
